+++ R/helpers.R
@@ -363,15 +363,15 @@ wls <- function(X, Y, w = NULL, return_se = FALSE) {
 #'
 #' form <- Y - f ~ X1
 #'
-#' X_l <- model.matrix(form, data = dat[dat$set == "labeled",])
+#' X_l <- model.matrix(form, data = dat[dat$set_label == "labeled",])
 #'
-#' Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)
+#' Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)
 #'
-#' f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)
+#' f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)
 #'
-#' X_u <- model.matrix(form, data = dat[dat$set == "unlabeled",])
+#' X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled",])
 #'
-#' f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)
+#' f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)
 #'
 #' est <- ppi_plusplus_ols_est(X_l, Y_l, f_l, X_u, f_u)
 #'
@@ -471,15 +471,15 @@ ols_get_stats <- function(est, X_l, Y_l, f_l, X_u, f_u,
 #'
 #' form <- Y - f ~ X1
 #'
-#' X_l <- model.matrix(form, data = dat[dat$set == "labeled",])
+#' X_l <- model.matrix(form, data = dat[dat$set_label == "labeled",])
 #'
-#' Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)
+#' Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)
 #'
-#' f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)
+#' f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)
 #'
-#' X_u <- model.matrix(form, data = dat[dat$set == "unlabeled",])
+#' X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled",])
 #'
-#' f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)
+#' f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)
 #'
 #' est <- ppi_plusplus_ols_est(X_l, Y_l, f_l, X_u, f_u)
 #'
@@ -778,15 +778,15 @@ log1pexp <- function(x) {
 #'
 #' form <- Y - f ~ X1
 #'
-#' X_l <- model.matrix(form, data = dat[dat$set == "labeled",])
+#' X_l <- model.matrix(form, data = dat[dat$set_label == "labeled",])
 #'
-#' Y_l <- dat[dat$set == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)
+#' Y_l <- dat[dat$set_label == "labeled", all.vars(form)[1]] |> matrix(ncol = 1)
 #'
-#' f_l <- dat[dat$set == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)
+#' f_l <- dat[dat$set_label == "labeled", all.vars(form)[2]] |> matrix(ncol = 1)
 #'
-#' X_u <- model.matrix(form, data = dat[dat$set == "unlabeled",])
+#' X_u <- model.matrix(form, data = dat[dat$set_label == "unlabeled",])
 #'
-#' f_u <- dat[dat$set == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)
+#' f_u <- dat[dat$set_label == "unlabeled", all.vars(form)[2]] |> matrix(ncol = 1)
 #'
 #' est <- ppi_plusplus_logistic_est(X_l, Y_l, f_l, X_u, f_u)
 #'