+++ R/ipd.R
@@ -70,7 +70,7 @@
 #'
 #' @param ... Additional arguments to be passed to the fitting function. See
 #' the \code{Details} section for more information. See
-#' \strong{5. Auxilliary Arguments} and \strong{6. Other Arguments} in the
+#' \strong{5. Auxiliary Arguments} and \strong{6. Other Arguments} in the
 #' \strong{Details} below for more information.
 #'
 #' @returns a summary of model output.
@@ -99,8 +99,8 @@
 #' stacked \code{data.frame} with both the unlabeled and labeled data and a
 #' \code{label} argument that specifies the column identifying the labeled
 #' versus the unlabeled observations in the stacked \code{data.frame} (e.g.,
-#' \code{label = "set"} if the column "set" in the stacked data denotes which
-#' set an observation belongs to).
+#' \code{label = "set_label"} if the column "set_label" in the stacked data
+#' denotes which set an observation belongs to).
 #'
 #' NOTE: Labeled data identifiers can be:
 #'
@@ -204,7 +204,7 @@
 #'
 #' ipd(formula, method = "postpi_analytic", model = "ols",
 #'
-#'     data = dat, label = "set")
+#'     data = dat, label = "set_label")
 #'
 #' #-- PostPI Bootstrap Correction (Wang et al., 2020)
 #'
@@ -212,25 +212,25 @@
 #'
 #' ipd(formula, method = "postpi_boot", model = "ols",
 #'
-#'     data = dat, label = "set", nboot = nboot)
+#'     data = dat, label = "set_label", nboot = nboot)
 #'
 #' #-- PPI (Angelopoulos et al., 2023)
 #'
 #' ipd(formula, method = "ppi", model = "ols",
 #'
-#'     data = dat, label = "set")
+#'     data = dat, label = "set_label")
 #'
 #' #-- PPI++ (Angelopoulos et al., 2023)
 #'
 #' ipd(formula, method = "ppi_plusplus", model = "ols",
 #'
-#'     data = dat, label = "set")
+#'     data = dat, label = "set_label")
 #'
 #' #-- PSPA (Miao et al., 2023)
 #'
 #' ipd(formula, method = "pspa", model = "ols",
 #'
-#'     data = dat, label = "set")
+#'     data = dat, label = "set_label")
 #'
 #' @import stats
 #'